# contracts/openapi/pomi-api-v1.yaml
openapi: 3.1.0
info:
  title: Pomi API
  version: 0.1.0
  description: |
    Pomi P0 Flutter/FastAPI contract. All medical data in the hackathon build is
    simulated. OCR output remains a draft until explicitly confirmed by the user.
    Hospital certification is a Flutter-only local demo and is intentionally absent.
    Authentication and health endpoints are already implemented and therefore use
    their direct response models. The remaining business endpoints are the agreed
    P0 contract and use the ApiSuccess/ApiError envelope until implemented.
servers:
  - url: https://api.healy1012-ops.top
    description: Demo server
  - url: http://127.0.0.1:8000
    description: Local development
security:
  - SessionBearer: []
...
paths:
...
  /api/ocr/tasks:
    post:
      tags: [OCR]
      operationId: createOcrTask
      summary: Queue OCR for a specific document revision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OcrTaskCreate'
      responses:
        '201':
          description: OCR task queued or existing idempotent task returned
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/OcrTask'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
  /api/ocr/tasks/{task_id}:
    get:
      tags: [OCR]
      operationId: getOcrTask
      summary: Poll OCR task status
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: OCR task
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/OcrTask'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /api/ocr/tasks/{task_id}/result:
    get:
      tags: [OCR]
      operationId: getOcrResult
      summary: Get the editable OCR draft and field-level evidence
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: OCR draft
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/OcrDraftResult'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /api/ocr/tasks/{task_id}/confirm:
    post:
      tags: [OCR]
      operationId: confirmOcrResult
      summary: Atomically confirm a laboratory or clinical-text OCR draft
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/LabConfirmRequest'
                - $ref: '#/components/schemas/ClinicalTextConfirmRequest'
                - $ref: '#/components/schemas/MedicalOrderConfirmRequest'
      responses:
        '200':
          description: OCR result confirmed and formal records created atomically
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        oneOf:
                          - $ref: '#/components/schemas/LabConfirmResult'
                          - $ref: '#/components/schemas/ClinicalTextConfirmResult'
                          - $ref: '#/components/schemas/MedicalOrderConfirmationResult'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
  /api/ocr/tasks/{task_id}/retry:
    post:
      tags: [OCR]
      operationId: retryOcrTask
      summary: Create a new user-requested OCR attempt
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OcrRetryRequest'
      responses:
        '201':
          description: New OCR attempt queued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/OcrTask'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
...
  /api/patient-notes:
    post:
      tags: [Reports]
      operationId: createPatientNote
      summary: Create a patient-note draft for the current visit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientNoteInput'
      responses:
        '201':
          description: Patient note created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PatientNote'
...
  /api/patient-notes/{note_id}:
    put:
      tags: [Reports]
      operationId: updatePatientNote
      summary: Update an unconsumed patient-note draft
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientNoteInput'
      responses:
        '200':
          description: Patient note updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PatientNote'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /api/patient-notes/{note_id}/confirm:
    post:
      tags: [Reports]
      operationId: confirmPatientNote
      summary: Idempotently confirm a patient note
      parameters:
        - $ref: '#/components/parameters/PatientNoteId'
      responses:
        '200':
          description: Patient note confirmed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PatientNote'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /api/patient-notes/{note_id}/skip:
    post:
      tags: [Reports]
      operationId: skipPatientNote
      summary: Idempotently skip a patient note
      parameters:
        - $ref: '#/components/parameters/PatientNoteId'
      responses:
        '200':
          description: Patient note skipped
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PatientNote'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /api/patient-notes/{note_id}/copy:
    post:
      tags: [Reports]
      operationId: copyPatientNote
      summary: Copy a prior patient note into a new editable draft
      parameters:
        - $ref: '#/components/parameters/PatientNoteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientNoteCopy'
      responses:
        '201':
          description: New patient-note draft created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PatientNote'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    SessionBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-id
  parameters:
...
    PatientNoteId:
      name: note_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ErrorResponse:
      description: API error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
    ApiSuccess:
      type: object
      required: [success, data, request_id, error]
      properties:
        success:
          type: boolean
          const: true
        data: {}
        request_id:
          type: string
        error:
          type: 'null'
    ApiError:
      type: object
      required: [success, data, request_id, error]
      properties:
        success:
          type: boolean
          const: false
        data:
          type: 'null'
        request_id:
          type: string
        error:
          $ref: '#/components/schemas/ErrorDetail'
...
    DocumentType:
      type: string
      enum: [lab_report, medical_order, imaging_text_report, outpatient_record]
...
    OcrTaskStatus:
      type: string
      enum: [queued, processing, pending_confirmation, confirmed, failed, timed_out]
    OcrResultSource:
      type: string
      enum: [qwen3-vl]
    OcrTaskCreate:
      type: object
      required: [document_id, document_revision_id]
      properties:
        document_id:
          type: string
          format: uuid
        document_revision_id:
          type: string
          format: uuid
    OcrTask:
      type: object
      required: [id, document_id, document_revision_id, material_type, status, model, prompt_version, schema_version, attempt_number, provider_attempts, attempt_history, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        document_revision_id:
          type: string
          format: uuid
        material_type:
          $ref: '#/components/schemas/DocumentType'
        status:
          $ref: '#/components/schemas/OcrTaskStatus'
        model:
          type: string
        prompt_version:
          type: string
        schema_version:
          type: string
        attempt_number:
          type: integer
          minimum: 1
        provider_attempts:
          type: integer
          minimum: 0
        parent_task_id:
          type: [string, 'null']
          format: uuid
        attempt_history:
          type: array
          items:
            type: object
            additionalProperties: true
        duration_ms:
          type: [integer, 'null']
        error:
          type: [object, 'null']
          additionalProperties: true
        result_source:
          oneOf:
            - $ref: '#/components/schemas/OcrResultSource'
            - type: 'null'
        reused:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    FieldConfirmationStatus:
      type: string
      enum: [pending, confirmed, edited, rejected]
    OcrFieldResult:
      type: object
      required: [id, path, source_text, parsed_value, confidence, confirmation_status]
      properties:
        id:
          type: string
          format: uuid
        path:
          type: string
        source_text:
          type: [string, 'null']
        parsed_value: {}
        confidence:
          type: [number, 'null']
          minimum: 0
          maximum: 1
        uncertainty_reason:
          type: [string, 'null']
        source_region:
          type: [object, 'null']
          additionalProperties: true
        user_value: {}
        confirmation_status:
          $ref: '#/components/schemas/FieldConfirmationStatus'
...
    LabDraft:
      type: object
      required: [hospital_name, sample_date, report_date, items]
      additionalProperties: false
      properties:
        hospital_name:
          type: [string, 'null']
        sample_date:
          type: [string, 'null']
          format: date
        report_date:
          type: [string, 'null']
          format: date
        items:
          type: array
          items:
            $ref: '#/components/schemas/LabItemDraft'
...
    MedicalOrderDraft:
      type: object
      required: [hospital_name, department_name, prescribed_at, orders]
      additionalProperties: false
      properties:
        hospital_name:
          type: [string, 'null']
        department_name:
          type: [string, 'null']
        prescribed_at:
          type: [string, 'null']
          format: date
        orders:
          type: array
          items:
            $ref: '#/components/schemas/MedicalOrderItemDraft'
    ImagingTextDraft:
      type: object
      required: [examination_name, body_part, examination_method, findings_text, conclusion_text, examined_at, reported_at]
      additionalProperties: false
      properties:
        examination_name:
          type: [string, 'null']
        body_part:
          type: [string, 'null']
        examination_method:
          type: [string, 'null']
        findings_text:
          type: [string, 'null']
        conclusion_text:
          type: [string, 'null']
        examined_at:
          type: [string, 'null']
          format: date
        reported_at:
          type: [string, 'null']
          format: date
    OutpatientRecordDraft:
      type: object
      required: [hospital_name, department_name, doctor_name, visit_date, chief_complaint, diagnosis_summary, treatment_plan, medical_advice]
      additionalProperties: false
      properties:
        hospital_name:
          type: [string, 'null']
        department_name:
          type: [string, 'null']
        doctor_name:
          type: [string, 'null']
        visit_date:
          type: [string, 'null']
          format: date
        chief_complaint:
          type: [string, 'null']
        diagnosis_summary:
          type: [string, 'null']
        treatment_plan:
          type: [string, 'null']
        medical_advice:
          type: [string, 'null']
    OcrDraftResult:
      type: object
      required: [id, task_id, raw_response, validated_draft, fields, created_at]
      properties:
        id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        raw_response:
          type: object
          additionalProperties: true
        validated_draft:
          oneOf:
            - $ref: '#/components/schemas/LabDraft'
            - $ref: '#/components/schemas/MedicalOrderDraft'
            - $ref: '#/components/schemas/ImagingTextDraft'
            - $ref: '#/components/schemas/OutpatientRecordDraft'
        user_modified_data:
          type: [object, 'null']
          additionalProperties: true
        confirmed_data:
          type: [object, 'null']
          additionalProperties: true
        fields:
          type: array
          items:
            $ref: '#/components/schemas/OcrFieldResult'
        source_document:
          $ref: '#/components/schemas/OcrSourceDocument'
        created_at:
          type: string
          format: date-time
    MedicalOrderConfirmationItem:
      type: object
      additionalProperties: false
      required: [index, confirmed, drug_name, dosage_value, dosage_unit, frequency, order_date, raw_order_text]
      properties:
        index:
          type: integer
          minimum: 0
        confirmed:
          type: boolean
          const: true
        drug_name:
          type: string
          minLength: 1
        specification:
          type: [string, 'null']
        dosage_value:
          type: number
          exclusiveMinimum: 0
        dosage_unit:
          type: string
          minLength: 1
        frequency:
          type: string
          minLength: 1
        course:
          type: [string, 'null']
        route:
          type: [string, 'null']
        instructions:
          type: [string, 'null']
        order_date:
          type: string
          format: date
        raw_order_text:
          type: string
          minLength: 1
        explicitly_stopped:
          type: boolean
          default: false
    MedicalOrderConfirmRequest:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/MedicalOrderConfirmationItem'
    MedicalOrderConfirmationResult:
      type: object
      required: [items, reused]
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
        reused:
          type: boolean
    OcrSourceDocument:
      type: object
      additionalProperties: false
      required: [document_id, document_revision_id, original_file_name, mime_type, revision_number, file_endpoint]
      properties:
        document_id:
          type: string
          format: uuid
        document_revision_id:
          type: string
          format: uuid
        original_file_name:
          type: string
        mime_type:
          type: string
          enum: [image/jpeg, image/png, application/pdf]
        revision_number:
          type: integer
          minimum: 1
        file_endpoint:
          type: string
          pattern: '^/api/documents/.+/revisions/.+/file$'
    ClinicalFieldConfirmation:
      type: object
      additionalProperties: false
      required: [field_path, user_value, confirmation_status]
      properties:
        field_path:
          type: string
          minLength: 1
          maxLength: 300
        user_value: {}
        confirmation_status:
          type: string
          enum: [confirmed, edited, rejected]
    ImagingTextConfirmation:
      type: object
      additionalProperties: false
      required: [findings_text, conclusion_text]
      properties:
        examination_name:
          type: [string, 'null']
          maxLength: 200
        body_part:
          type: [string, 'null']
          maxLength: 200
        examination_method:
          type: [string, 'null']
          maxLength: 200
        findings_text:
          type: string
          minLength: 1
          maxLength: 20000
        conclusion_text:
          type: string
          minLength: 1
          maxLength: 20000
        examined_at:
          type: [string, 'null']
          format: date
        reported_at:
          type: [string, 'null']
          format: date
    OutpatientRecordConfirmation:
      type: object
      additionalProperties: false
      required: [visit_date, diagnosis_summary, medical_advice]
      properties:
        hospital_name:
          type: [string, 'null']
          maxLength: 200
        department_name:
          type: [string, 'null']
          maxLength: 200
        doctor_name:
          type: [string, 'null']
          maxLength: 100
        visit_date:
          type: string
          format: date
        chief_complaint:
          type: [string, 'null']
          maxLength: 20000
        diagnosis_summary:
          type: string
          minLength: 1
          maxLength: 20000
        treatment_plan:
          type: [string, 'null']
          maxLength: 20000
        medical_advice:
          type: string
          minLength: 1
          maxLength: 20000
    ClinicalTextConfirmRequest:
      type: object
      additionalProperties: false
      required: [result_id, expected_revision_id, document_type, confirmed_data]
      properties:
        result_id:
          type: string
          format: uuid
        expected_revision_id:
          type: string
          format: uuid
        document_type:
          type: string
          enum: [imaging_text_report, outpatient_record]
        confirmed_data:
          oneOf:
            - $ref: '#/components/schemas/ImagingTextConfirmation'
            - $ref: '#/components/schemas/OutpatientRecordConfirmation'
        field_confirmations:
          type: array
          items:
            $ref: '#/components/schemas/ClinicalFieldConfirmation'
          default: []
        confirm_all:
          type: boolean
          default: false
...
    ClinicalTextConfirmResult:
      type: object
      required: [record_id, task_id, result_id, created_resource_ids, reconciliation_required, material_type, document_id, document_revision_id, ocr_result_id, confirmed_at, summary, p0_evaluation, reused]
      properties:
        record_id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        result_id:
          type: string
          format: uuid
        created_resource_ids:
          type: array
          items:
            type: string
            format: uuid
        reconciliation_required:
          type: boolean
          const: false
        material_type:
          type: string
          enum: [imaging_text_report, outpatient_record]
        document_id:
          type: string
          format: uuid
        document_revision_id:
          type: string
          format: uuid
        ocr_result_id:
          type: string
          format: uuid
        confirmed_at:
          type: string
          format: date-time
        summary:
          type: object
          additionalProperties: true
        p0_evaluation:
          $ref: '#/components/schemas/ClinicalTextP0Evaluation'
        reused:
          type: boolean
    LabConfirmationItem:
      type: object
      additionalProperties: false
      required: [name, value, unit]
      properties:
        source_index:
          type: [integer, 'null']
          minimum: 0
        name:
          type: [string, 'null']
          maxLength: 200
        value:
          type: [string, number, 'null']
        unit:
          type: [string, 'null']
          maxLength: 40
        reference_range:
          type: [string, 'null']
          maxLength: 120
        sample_date:
          type: [string, 'null']
        exam_date:
          type: [string, 'null']
        report_date:
          type: [string, 'null']
        visit_date:
          type: [string, 'null']
        note:
          type: [string, 'null']
          maxLength: 1000
    LabConfirmRequest:
      type: object
      additionalProperties: false
      required: [result_id, expected_revision_id, items]
      properties:
        result_id:
          type: string
          format: uuid
        expected_revision_id:
          type: string
          format: uuid
        visit_id:
          type: [string, 'null']
          maxLength: 100
        sample_date:
          type: [string, 'null']
        exam_date:
          type: [string, 'null']
        report_date:
          type: [string, 'null']
        visit_date:
          type: [string, 'null']
        items:
          type: array
          maxItems: 200
          items:
            $ref: '#/components/schemas/LabConfirmationItem'
    LabP0Evaluation:
      type: object
      required: [total_fields, valid_fields, invalid_fields, valid_rate]
      properties:
        total_fields:
          type: integer
          minimum: 0
        valid_fields:
          type: integer
          minimum: 0
        invalid_fields:
          type: integer
          minimum: 0
        valid_rate:
          type: number
          minimum: 0
          maximum: 1
        ocr_exact_match_fields:
          type: integer
          minimum: 0
        user_corrected_fields:
          type: integer
          minimum: 0
        ocr_exact_match_rate:
          type: number
          minimum: 0
          maximum: 1
    LabObservation:
      type: object
      required: [id, patient_id, document_id, document_revision_id, ocr_result_id, item_index, original_item_name, mapping_status, raw_value, numeric_value, original_unit, standard_unit, abnormal_status, original_item_data, confirmed_item_data, confirmed_by_uid, confirmed_at]
      properties:
        id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        visit_id:
          type: [string, 'null']
        document_id:
          type: string
          format: uuid
        document_revision_id:
          type: string
          format: uuid
        ocr_result_id:
          type: string
          format: uuid
        item_index:
          type: integer
          minimum: 0
        original_item_name:
          type: string
        standard_metric_id:
          type: [string, 'null']
        mapping_status:
          type: string
          enum: [mapped, needs_manual_review]
        raw_value:
          type: string
        numeric_value:
          type: string
        original_unit:
          type: string
        standard_unit:
          type: string
        reference_range_raw:
          type: [string, 'null']
        reference_lower:
          type: [string, 'null']
        reference_upper:
          type: [string, 'null']
        abnormal_status:
          type: string
          enum: [low, normal, high, unknown]
        sample_date:
          type: [string, 'null']
          format: date
        exam_date:
          type: [string, 'null']
          format: date
        report_date:
          type: [string, 'null']
          format: date
        visit_date:
          type: [string, 'null']
          format: date
        trend_date:
          type: [string, 'null']
          format: date
        trend_date_source:
          type: [string, 'null']
          enum: [sample_date, exam_date, report_date, visit_date, null]
        original_item_data:
          type: object
          additionalProperties: true
        confirmed_item_data:
          type: object
          additionalProperties: true
        confirmed_by_uid:
          type: string
          format: uuid
        confirmed_at:
          type: string
          format: date-time
        note:
          type: [string, 'null']
    LabConfirmResult:
      type: object
      required: [task_id, result_id, status, reused, created_resource_ids, confirmed_at, observations, p0_evaluation]
      properties:
        task_id:
          type: string
          format: uuid
        result_id:
          type: string
          format: uuid
        status:
          type: string
          const: confirmed
        reused:
          type: boolean
        created_resource_ids:
          type: array
          items:
            type: string
            format: uuid
        confirmed_at:
          type: string
          format: date-time
        observations:
          type: array
          items:
            $ref: '#/components/schemas/LabObservation'
        p0_evaluation:
          $ref: '#/components/schemas/LabP0Evaluation'
    OcrRetryRequest:
      type: object
      properties:
        reason:
          type: [string, 'null']
          maxLength: 500
        allow_registered_fallback:
          type: boolean
          default: true
...
    PatientNoteStatus:
      type: string
      enum: [draft, confirmed, skipped, consumed]
    PatientNoteInput:
      type: object
      required: [original_text]
      additionalProperties: false
      properties:
        original_text:
          type: string
          maxLength: 5000
        visit_context:
          type: [string, 'null']
          maxLength: 200
    PatientNoteCopy:
      type: object
      properties:
        visit_context:
          type: [string, 'null']
          maxLength: 200
    PatientNote:
      type: object
      required: [id, patient_id, original_text, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        visit_context:
          type: [string, 'null']
        original_text:
          type: string
        confirmed_text:
          type: [string, 'null']
        status:
          $ref: '#/components/schemas/PatientNoteStatus'
        source_note_id:
          type: [string, 'null']
          format: uuid
        confirmed_by_uid:
          type: [string, 'null']
          format: uuid
        confirmed_at:
          type: [string, 'null']
          format: date-time
        consumed_at:
          type: [string, 'null']
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...